asiopq::connection::update_socket Improvement
Improved the method by which asiopq::connection::update_socket keeps the ASIO socket up to date.

Previously asiopq::connection::update_socket used a different method between Windows and POSIX.  On Windows it would use WSADuplicateSocketW to retrieve a WSAPROTOCOL_INFOW structure.  This structure contains a dwCatalogEntryId which MSDN describes as "[a] unique identifier assigned by the WS2_32.DLL for each WSAPROTOCOL_INFO structure".  This was compared against a cached value from the last time WSADuplicateSocketW was called and if the value was different the socket was duplicated via a call to WSASocketW, otherwise the old socket and new socket were the same and WSASocketW was not called.  On POSIX the socket was duplicated by calling dup each time asiopq::connection::update_socket was called, clearly not ideal.

The new method is based on the following observation: Since the socket has been duplicated it will remain bound to an endpoint until both copies are closed.  Given that we control one copy we know that when asiopq::connection::update_socket is executing the socket is bound to an endpoint.  Accordingly the operating system cannot reuse that endpoint.  Therefore we use getsockname and compare the resulting struct sockaddr_storage to a cache struct sockaddr_storage.  If these structures are byte-for-byte identical then the socket has not changed and no duplication occurs.  If they are not byte-for-byte identical the socket has changed and the new socket is duplicated by calling WSADuplicateSocketW followed by WSASocketW (on Windows) or by calling dup (on POSIX).