asiopq::connection::next - Error Handling Improvement
Previously asiopq::connection::next did two things which were degenerate in the face of errors/exceptions (specifically connect-/socket-related errors):

1.	It cleared the asiopq::connection object's pending operation after performing socket operations (which could throw and thus prevent the reset from occurring)
2.	It unconditionally called asiopq::asio::tcp::ip::socket::cancel (this throws when the asiopq::asio::tcp::ip::socket does not wrap a valid socket descriptor)

This would have the effect of either leaving the asiopq::connection in an invalid/nonsensical state (in the former case), or throwing from the invoking asiopq::asio::io_service::run (in the latter case).  The correct outcome is to either leave the object in a state that makes some sense (in the former case), or to allow all pending operations to run (and fail since the libpq handle is clearly in an invalid state) (in the latter case).

Fixed this by:

1.	Reordering operations in asiopq::connection::reset such that the asiopq::connection object's pending operation will always be cleared
2.	Gating the call to asiopq::asio::tcp::ip::socket::cancel behind the return value of asiopq::asio::tcp::ip::socket::is_open

Added an integration test for this situation (i.e. failing connect with other pending operations).